Conflicts:
	include/text.php
	view/templates/head.tpl
	view/theme/duepuntozero/style.css
	view/theme/vier/style.css

+++ include/poller.php
@@ -26,17 +26,11 @@ function poller_run(&$argv, &$argc){
 		unset($db_host, $db_user, $db_pass, $db_data);
 	};
 
-	$load = current_load();
-	if($load) {
-		$maxsysload = intval(get_config('system','maxloadavg'));
-		if($maxsysload < 1)
-			$maxsysload = 50;
+	if (poller_max_connections_reached())
+		return;
 
-		if(intval($load) > $maxsysload) {
-			logger('system: load ' . $load . ' too high. poller deferred to next scheduled run.');
-			return;
-		}
-	}
+	if (App::maxload_reached())
+		return;
 
 	// Checking the number of workers
 	if (poller_too_much_workers(1)) {
@@ -64,6 +58,10 @@ function poller_run(&$argv, &$argc){
 	$starttime = time();
 
 	while ($r = q("SELECT * FROM `workerqueue` WHERE `executed` = '0000-00-00 00:00:00' ORDER BY `created` LIMIT 1")) {
+
+		// Constantly check the number of available database connections to let the frontend be accessible at any time
+		if (poller_max_connections_reached())
+			return;
 
 		// Count active workers and compare them with a maximum value that depends on the load
 		if (poller_too_much_workers(3))
@@ -118,11 +116,92 @@ function poller_run(&$argv, &$argc){
 }
 
 /**
+ * @brief Checks if the number of database connections has reached a critical limit.
+ *
+ * @return bool Are more than 3/4 of the maximum connections used?
+ */
+function poller_max_connections_reached() {
+
+	// Fetch the max value from the config. This is needed when the system cannot detect the correct value by itself.
+	$max = get_config("system", "max_connections");
+
+	if ($max == 0) {
+		// the maximum number of possible user connections can be a system variable
+		$r = q("SHOW VARIABLES WHERE `variable_name` = 'max_user_connections'");
+		if ($r)
+			$max = $r[0]["Value"];
+
+		// Or it can be granted. This overrides the system variable
+		$r = q("SHOW GRANTS");
+		if ($r)
+			foreach ($r AS $grants) {
+				$grant = array_pop($grants);
+				if (stristr($grant, "GRANT USAGE ON"))
+					if (preg_match("/WITH MAX_USER_CONNECTIONS (\d*)/", $grant, $match))
+						$max = $match[1];
+			}
+	}
+
+	// If $max is set we will use the processlist to determine the current number of connections
+	// The processlist only shows entries of the current user
+	if ($max != 0) {
+		$r = q("SHOW PROCESSLIST");
+		if (!$r)
+			return false;
+
+		$used = count($r);
+
+		logger("Connection usage (user values): ".$used."/".$max, LOGGER_DEBUG);
+
+		$level = $used / $max;
+
+		if ($level >= (3/4)) {
+			logger("Maximum level (3/4) of user connections reached: ".$used."/".$max);
+			return true;
+		}
+	}
+
+	// We will now check for the system values.
+	// This limit could be reached although the user limits are fine.
+	$r = q("SHOW VARIABLES WHERE `variable_name` = 'max_connections'");
+	if (!$r)
+		return false;
+
+	$max = intval($r[0]["Value"]);
+	if ($max == 0)
+		return false;
+
+	$r = q("SHOW STATUS WHERE `variable_name` = 'Threads_connected'");
+	if (!$r)
+		return false;
+
+	$used = intval($r[0]["Value"]);
+	if ($used == 0)
+		return false;
+
+	logger("Connection usage (system values): ".$used."/".$max, LOGGER_DEBUG);
+
+	$level = $used / $max;
+
+	if ($level < (3/4))
+		return false;
+
+	logger("Maximum level (3/4) of system connections reached: ".$used."/".$max);
+	return true;
+}
+
+/**
  * @brief fix the queue entry if the worker process died
  *
  */
 function poller_kill_stale_workers() {
 	$r = q("SELECT `pid`, `executed` FROM `workerqueue` WHERE `executed` != '0000-00-00 00:00:00'");
+
+	if (!is_array($r) || count($r) == 0) {
+		// No processing here needed
+		return;
+	}
+
 	foreach($r AS $pid)
 		if (!posix_kill($pid["pid"], 0))
 			q("UPDATE `workerqueue` SET `executed` = '0000-00-00 00:00:00', `pid` = 0 WHERE `pid` = %d",
